Corrects two Cartridge type mismatches.

+++ Storage/Cartridge/Cartridge.hpp
@@ -31,7 +31,7 @@ class Cartridge {
 			Segment(size_t start_address, size_t end_address, std::vector<uint8_t> data) :
 				start_address(start_address), end_address(end_address), data(std::move(data)) {}
 
-			Segment(int start_address, std::vector<uint8_t> data) :
+			Segment(size_t start_address, std::vector<uint8_t> data) :
 				Segment(start_address, start_address + data.size(), data) {}
 
 			/// Indicates that an address is unknown.
